pyproject.toml
[tool.poetry]
name = "metadoc"
version = "0.1.0"
description = "Document Defined Agent"
authors = ["Leric <leric.zhang@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airic"}]

[tool.poetry.dependencies]
python = "^3.12"
google-adk = "^0.1.0"
litellm = "^1.66.3"
typer = {extras = ["all"], version = "^0.9.0"}
python-frontmatter = "^1.1.0"
markdown2 = "^2.4.12"
prompt-toolkit = "^3.0.43"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
airic = "airic.cli.main:main"
